requirements.txt
matplotlib>=3.5.3
python-dotenv>=1.0.1
python-telegram-bot>=21.1.1
python-telegram-bot[job-queue]
yfinance>=0.2.38